feat: Implement builtin of exit
