+++ include/builtins.h
@@ -28,4 +28,6 @@ int		ft_exec_echo(char **args);
 int		ft_exec_pwd(void);
 void	exec_exit(t_shell *shell, t_final_cmd_table *final_cmd_table);
 
+int		get_args_error(char **args);
+
 #endif
